test(github): split format tests by concern
github/format/tests.rs (586) split into four sibling files: size_tests,
tree_tests, overview_tests (includes readme truncation + the shared
sample_repo/make_readme_bytes/parse_shown_bytes helpers), file_content_tests.
All <=400 lines; production format.rs unchanged apart from mod declarations;
449 tests pass.
